dbengine: carve storage-engine-types.h, decouple from rrd.h (SOW-0018)
Header reorganization that lets external consumers include
src/database/engine/rrdengine.h without dragging in the whole daemon
header graph (streaming, sqlite, ACLK, contexts, health, web/api).
First step in the dbengine library extraction series tracked by
.agents/sow/specs/dbengine-library.md.

Changes:

- New header src/database/storage-engine-types.h holds the POD type
  and enum surface needed by the dbengine: STORAGE_PRIORITY,
  STORAGE_ENGINE_BACKEND + is_valid_backend, storage_engine_query_handle,
  STORAGE_INSTANCE / STORAGE_METRIC_HANDLE / STORAGE_METRICS_GROUP
  opaque typedefs, STORAGE_COLLECT_HANDLE, RRDDIM / RRDSET forward
  decls, and RRD_BACKFILL (moved here from rrddim-backfill.h). The
  header depends only on libnetdata and rrd-database-mode.h.

- src/database/storage-engine.h now includes storage-engine-types.h
  and drops the duplicated definitions; src/database/rrddim-backfill.h
  re-exports RRD_BACKFILL through the new header.

- src/database/engine/rrdengine.h, mrg.h, and cache.h now include
  storage-engine-types.h instead of the kitchen-sink ../rrd.h.

- protected-access moves from src/daemon/protected-access.{c,h} to
  src/libnetdata/protected-access/protected-access.{c,h}. The
  rrdengine.h include of "daemon/protected-access.h" is removed;
  consumers reach it transitively through libnetdata.h. The
  src/daemon/signal-handler.c explicit include is also dropped.

- CMakeLists.txt: protected-access files move from DAEMON_FILES to
  LIBNETDATA_FILES; storage-engine-types.h is added to RRD_PLUGIN_FILES.

- cache.h adds a forward declaration of netdata_conf_cpus() so its
  inline functions still compile after losing the transitive rrd.h
  include; SOW-0021 will route this through the library config struct.

- The 12 dbengine .c files that previously got daemon symbols
  (pulse_*, UV_EVENT_DBENGINE_*, populate_metrics_from_database,
  nd_profile, localhost, RRDSET, RRDDIM) through ../rrd.h via the
  engine .h chain now include ../rrd.h explicitly. SOWs 0020-0022 will
  progressively eliminate these .c-side daemon dependencies by
  routing them through the library configuration and pulse
  callbacks.

- src/daemon/pulse/pulse-db-dbengine-retention.c adds an explicit
  include of database/rrd.h that it previously got transitively via
  the now-cleaned engine API header.

Verification:

- just cfg && just ninja completes through [189/189] Linking CXX
  executable netdata with no new diagnostics.

- A stub TU including only rrdengine.h, preprocessed with clang -E,
  pulls in 142 headers total. Zero of them match
  streaming|sqlite|aclk|/contexts/|/web/|/health/|/rrdhost\.h|
  /rrdset\.h|/rrddim\.h. Zero top-level src/database/*.h headers are
  pulled in.

SOW closure: .agents/sow/pending/SOW-0018-... moves to
.agents/sow/done/SOW-0018-... with Status: completed in this commit.
