+++ src/database/engine/cache.h
@@ -3,7 +3,12 @@
 #define DBENGINE_CACHE_H
 
 #include "datafile.h"
-#include "../rrd.h"
+#include "../storage-engine-types.h"
+
+// Forward declaration of daemon-side helper; the full declaration lives
+// in src/daemon/config/netdata-conf-global.h. SOW-0021 will route this
+// through the library configuration.
+size_t netdata_conf_cpus(void);
 
 // CACHE COMPILE TIME CONFIGURATION
 // #define PGC_COUNT_POINTS_COLLECTED 1
