+++ src/database/engine/dbengine-compression.c
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+#include "../rrd.h"
 #include "rrdengine.h"
 #include "dbengine-compression.h"
