+++ src/database/engine/mrg-unittest.c
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+#include "../rrd.h"
 #include "mrg-internals.h"
 #include "rrdengine.h"
