+++ src/database/engine/mrg.h
@@ -2,7 +2,7 @@
 #ifndef DBENGINE_METRIC_H
 #define DBENGINE_METRIC_H
 
-#include "../rrd.h"
+#include "../storage-engine-types.h"
 
 typedef struct metric METRIC;
 typedef struct mrg MRG;
