+++ src/database/engine/page.c
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+#include "../rrd.h"
 #include "page.h"
 
 #include "libnetdata/libnetdata.h"
