+++ src/database/engine/pagecache.c
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+#include "../rrd.h"
 #include "rrdengine.h"
 
 MRG *main_mrg = NULL;
