+++ src/database/engine/pdc.c
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+#include "../rrd.h"
 #include "pdc.h"
 #include "dbengine-compression.h"
