+++ src/database/engine/rrdengineapi.c
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+#include "../rrd.h"
 #include "database/engine/rrddiskprotocol.h"
 #include "rrdengine.h"
 #include "dbengine-compression.h"
