+++ src/database/rrddim-backfill.h
@@ -3,12 +3,7 @@
 #ifndef NETDATA_RRDDIM_BACKFILL_H
 #define NETDATA_RRDDIM_BACKFILL_H
 
-typedef enum __attribute__ ((__packed__)) {
-    RRD_BACKFILL_NONE = 0,
-    RRD_BACKFILL_FULL,
-    RRD_BACKFILL_NEW
-} RRD_BACKFILL;
-
+#include "storage-engine-types.h"
 #include "rrddim.h"
 
 bool backfill_tier_from_smaller_tiers(RRDDIM *rd, size_t tier, time_t now_s);
