+++ src/libnetdata/protected-access/protected-access.c
